requirements.txt
anyio==3.6.2
APScheduler==3.6.3
black==23.1.0
cachetools==4.2.2
certifi==2022.12.7
click==8.1.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
mypy-extensions==0.4.3
numpy==1.24.1
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
python-dotenv==0.21.1
python-telegram-bot==13.11
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
